Update doc links to use Yubico's new command names
Several command names changed recently (see #167).

This finishes out the name updates by fixing all documentation and links
to both docs.rs and the Yubico web site.
