+++ tests/command/blink_device.rs
@@ -1,6 +1,6 @@
 /// Blink the LED on the YubiHSM for 2 seconds
 #[test]
-fn blink_test() {
+fn blink_device_test() {
     let mut client = crate::get_hsm_client();
-    client.blink(2).unwrap();
+    client.blink_device(2).unwrap();
 }
